convert into spring boot and use new data generator
