+++ src/main/java/com/camunda/demo/customeronboarding/adapter/CalculateScore.java
@@ -4,17 +4,18 @@ import java.time.ZoneId;
 
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.JavaDelegate;
+import org.springframework.stereotype.Component;
 
 import com.camunda.demo.customeronboarding.model.NewApplication;
 
+@Component
 public class CalculateScore implements JavaDelegate {
-
 
 	@Override
 	public void execute(DelegateExecution execution) throws Exception {
 		NewApplication application = (NewApplication) execution.getVariable("application");
 		int yearLastDigit = application.getApplicant().getBirthday().toInstant().atZone(ZoneId.systemDefault()).toLocalDate().getYear() % 10;
-		
+
 		int score = 97;
 		
 		if(yearLastDigit == 3) {
