+++ src/main/java/com/camunda/demo/customeronboarding/facade/ApplicationOnlineFacade.java
@@ -4,34 +4,31 @@ import static org.camunda.spin.Spin.JSON;
 
 import java.io.UnsupportedEncodingException;
 
-import javax.ws.rs.Consumes;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.HttpHeaders;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
 import org.camunda.bpm.BpmPlatform;
 import org.camunda.bpm.engine.impl.digest._apacheCommonsCodec.Base64;
 import org.camunda.bpm.engine.runtime.ProcessInstance;
 import org.camunda.bpm.engine.variable.Variables;
 import org.camunda.bpm.engine.variable.Variables.SerializationDataFormats;
 import org.camunda.bpm.engine.variable.value.FileValue;
+import org.camunda.bpm.spring.boot.starter.webapp.CamundaBpmWebappInitializer;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import com.camunda.demo.customeronboarding.ProcessConstants;
 import com.camunda.demo.customeronboarding.model.NewApplication;
 
-@Path("/")
+@RestController
+@RequestMapping("API")
 public class ApplicationOnlineFacade {
 
-  @POST
-  @Path("new-application/{lang}")
-  @Consumes(MediaType.APPLICATION_JSON)
-  @Produces(MediaType.TEXT_HTML)
-  public Response submitNewApplication(NewApplication application, @Context HttpHeaders headers, @PathParam("lang") String lang) {
+  @PostMapping(path="/new-application/{lang}", produces=MediaType.TEXT_HTML_VALUE, consumes=MediaType.APPLICATION_JSON_VALUE)
+  public String submitNewApplication(@RequestBody NewApplication application, @RequestHeader("referer") String referer, @PathVariable("lang") String lang) {
     String processInstanceKey;
     if ("en".equals(lang)) {
       processInstanceKey = ProcessConstants.PROCESS_KEY_customer_onboarding_en;
@@ -40,10 +37,8 @@ public class ApplicationOnlineFacade {
     } else {
       throw new RuntimeException("Unsupported language requested.");
     }
-    String referer = headers.getHeaderString("Referer");
-    System.out.println("ich bin der refererer: " + referer);
     String uiBaseUrl = referer.substring(0, referer.lastIndexOf('/')) + "/";
-
+    
     // we start by using the shown price - user can change it on decision form
     application.setPremiumInCent(application.getPriceIndicationInCent());
 
@@ -57,13 +52,11 @@ public class ApplicationOnlineFacade {
             .putValue(ProcessConstants.VAR_NAME_applicantName, application.getApplicant().getName())//
             );
     
-    return Response.status(200).entity(processInstance.getBusinessKey()).build();
+    return processInstance.getBusinessKey();
   }
 
-  @POST
-  @Path("document/{number}")
-  @Consumes(MediaType.APPLICATION_JSON)
-  public void submitDocument(@PathParam("number") String number, VariableValueDto documentVariable) throws UnsupportedEncodingException {
+  @PostMapping(path="/document/{number}")
+  public void submitDocument(@PathVariable("number") String number, @RequestBody VariableValueDto documentVariable) throws UnsupportedEncodingException {
 
     FileValue document = Variables.fileValue((String) documentVariable.getValueInfo().get("filename"))
         .file(Base64.decodeBase64((String) documentVariable.getValue())) // see
