--- src/main/java/com/camunda/demo/customeronboarding/facade/RestApplication.java
@@ -1,10 +1,0 @@
-package com.camunda.demo.customeronboarding.facade;
-
-import javax.ws.rs.ApplicationPath;
-import javax.ws.rs.core.Application;
-
-@ApplicationPath("api")
-public class RestApplication extends Application {
-
-}
-
