+++ src/main/resources/META-INF/processes.xml
@@ -6,7 +6,7 @@
 		<process-engine>default</process-engine>
 		<properties>
 			<property name="isDeleteUponUndeploy">false</property>
-			<property name="isScanForProcessDefinitions">false</property>
+			<property name="isScanForProcessDefinitions">true</property>
 		</properties>
 	</process-archive>
