+++ src/main/resources/application.yml
@@ -1,0 +1,10 @@
+spring:
+ datasource:
+  url: 'jdbc:h2:~/testdb;DB_CLOSE_DELAY=-1;MVCC=TRUE'
+  username: sa
+  password: sa
+  driverClassName: 'org.h2.Driver'
+ 
+server:
+  servlet:
+    context-path: /camunda
