+++ src/main/webapp/online/insurance/app.js
@@ -38,7 +38,7 @@ $(document).ready(function() {
 
 
   var pathArray = window.location.pathname.split('/');
-  var baseUrl = window.location.protocol + "//" + window.location.host + "/" + pathArray[1] + "/api";
+  var baseUrl = window.location.protocol + "//" + window.location.host + "/" + pathArray[1] + "/API";
 
   $.i18n().load({
     'en': window.langEn['en'],
